envs/tbench2_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-tbench2_env"
version = "0.1.0"
description = "Tbench2 Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core>=0.2.0",
    "camel-ai",
    "pytest>=8.4.0",
    # Environment-specific dependencies
    # Add all dependencies needed for your environment here
    # Examples:
    # "numpy>=1.19.0",
    # "torch>=2.0.0",
    # "gymnasium>=0.29.0",
    # "openspiel>=1.0.0",
    # "smolagents>=1.22.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m tbench2_env.server.app
server = "tbench2_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["tbench2_env", "tbench2_env.server"]
package-dir = { "tbench2_env" = ".", "tbench2_env.server" = "server" }